+++ equalsverifier-core/src/test/java/nl/jqno/equalsverifier/internal/util/FormatterTest.java
@@ -62,7 +62,7 @@ public class FormatterTest {
             new NoFieldsAndThrowsNullMessage()
         );
         assertEquals(
-            "No fields, null message: [NoFieldsAndThrowsNullMessage]-throws NullPointerException(null)",
+            "No fields, null message: [NoFieldsAndThrowsNullMessage (no fields)]-throws NullPointerException(null)",
             f.format()
         );
     }
@@ -71,10 +71,7 @@ public class FormatterTest {
     public void oneAbstractParameter() {
         Instantiator<Abstract> i = Instantiator.of(Abstract.class);
         Formatter f = Formatter.of("Abstract: %%", i.instantiate());
-        assertThat(
-            f.format(),
-            containsString("Abstract: [Abstract x=0]-throws AbstractMethodError")
-        );
+        assertThat(f.format(), containsString("Abstract: [Abstract x=0]"));
     }
 
     @Test
@@ -88,10 +85,7 @@ public class FormatterTest {
     public void oneDelegatedAbstractParameter() {
         Instantiator<AbstractDelegation> i = Instantiator.of(AbstractDelegation.class);
         Formatter f = Formatter.of("Abstract: %%", i.instantiate());
-        assertThat(
-            f.format(),
-            containsString("Abstract: [AbstractDelegation y=0]-throws AbstractMethodError")
-        );
+        assertThat(f.format(), containsString("Abstract: [AbstractDelegation y=0]"));
     }
 
     @Test
@@ -117,16 +111,10 @@ public class FormatterTest {
         AbstractContainer ac = new AbstractContainer(i.instantiate());
 
         Formatter f = Formatter.of("AC: %%", ac);
-        String actual = f.format();
-        // Split up the message, because on some JDKs, AbstractMethodError is sometimes empty while
-        // on others, it isn't.
         assertThat(
-            actual,
-            containsString(
-                "AC: [AbstractContainer ad=[AbstractDelegation y=0]-throws AbstractMethodError"
-            )
+            f.format(),
+            containsString("AC: [AbstractContainer ad=[AbstractDelegation y=0]]")
         );
-        assertThat(actual, containsString("]-throws AbstractMethodError"));
     }
 
     @Test