+++ CHANGELOG.md
@@ -20,6 +20,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - Fix `ClassFormatError` when attempting to create a dynamic subclass for a class that has no package. ([Issue 638](https://github.com/jqno/equalsverifier/issues/638))
+- Fix `LinkageError` when running Quarkus dev mode and testing an abstract class. ([Issue 550](https://github.com/jqno/equalsverifier/issues/550))
 
 ## [3.12] - 2022-11-30