+++ CHANGELOG.md
@@ -13,6 +13,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [3.12.1] - 2022-12-02
+
 ### Changed
 
 - Improved error message output for objects that throw exceptions in their `toString()` methods. ([Issue 716](https://github.com/jqno/equalsverifier/issues/716))
@@ -1067,7 +1069,8 @@ Please don't use version 1.3; [it's a broken release](https://jqno.nl/post/2013/
 
 You can now use EqualsVerifier!
 
-[unreleased]: https://github.com/jqno/equalsverifier/compare/equalsverifier-3.12...HEAD
+[unreleased]: https://github.com/jqno/equalsverifier/compare/equalsverifier-3.12.1...HEAD
+[3.12.1]: https://github.com/jqno/equalsverifier/compare/equalsverifier-3.12...equalsverifier-3.12.1
 [3.12]: https://github.com/jqno/equalsverifier/compare/equalsverifier-3.11.1...equalsverifier-3.12
 [3.11.1]: https://github.com/jqno/equalsverifier/compare/equalsverifier-3.11...equalsverifier-3.11.1
 [3.11]: https://github.com/jqno/equalsverifier/compare/equalsverifier-3.10.1...equalsverifier-3.11